Add more unit tests for Basis creation function and overall matrix generation
Tests for short (5-element) bases in the k=1 subspace. Check that
elements are as expected and energies are right.

Tests for the Hamiltonian in the 5-element k=1 basis.

Update readme.

Add a __len__ magic function to Basis which does the sensible thing of returning
the length of the basis.

Continue adding docstrings to document the important classes and methods
for ease of reference later. To-do: set up automatic documentation?