Implement fermion states in qedstatefuncs
Add functions to compute energy and momentum from an occupation list in
statefuncs.py. These are currently unused, as it seems faster to create
temporary states for the purpose of building the basis. But if the state
creation is a limiting factor we can maybe just deal with the occupation
lists instead.

FermionState is now totally implemented in qedstatefuncs, as well as
FermionBasis and the functions to build a basis of fermion states. More
testing is needed to ensure that it works as expected but so far all
seems well. FermionState takes two occupation lists, one for the left
movers and one for the right movers. There is an optional flag to check
if the state is charge neutral. I don't yet have a way to generate only
the charge-neutral states; this may be a later step. But for now I
generate all states with zero momentum by construction and then get rid
of states with net charge.

Some additional test scripts are drafted in the tests folder.

To-do:
Continue testing FermionState and FermionBasis
Add a data or matrices folder and use datetime to store the matrices
generated by genmatrix with unique names for later reference
Implement the interaction term and begin generating matrices for the
QED states