Add massless spinor test, rename SpinorState args
The spinor wavefunctions for massless fermions have a simple
form. This adds a test to ensure that the u and v spinor
wavefunctions have the correct properties for values of momentum.

Note there is some conflict between what appears in Markus's
notes and the massless limit of the massive case (up to a
sign). Will check this before implementing next steps.

This commit also adds an extra timer statement to phi4_benchmark.

The naming convention for the inputs for SpinorState have also
been updated to clarify that u corresponds to particles and v