Implement interaction term for Schwinger model
The current implementation is a little inefficient. Currently we
loop over (\psi^\dagger \psi)_k (\psi^\dagger \psi)_{-k}, and there
is no shortcut to exclude operators which would lie outside the
energy range.

I am trying the approach of writing out all 16 operators in this
product and then normal-ordering them by hand.

Note that normal-ordering of fermionic operators has an ordering
ambiguity. That is, even after anticommuting all the annihilation
operators to the right, if we have e.g. a^\dagger_1 a^\dagger_2,
this operator is normal-ordered but so is a^\dagger_2 a^\dagger_1.

In particular the basis Fock states also suffer this ordering
(sign) ambiguity. In the next commit, we will chose a convention
for the ordering of operators in the normal-ordering and in the
definition of our Fock states, and update the implementation of
FermionOperator correspondingly.

This commit also adds the normed argument to the spinor wavefunctions
uspinor and vspinor which absorbs the 1/sqrt(2*omega).

The FermionBasis lookup function has been updated to drop parity
normalization factors of 1/sqrt(2) which were present in the phi4
theory, and the operator apply functions have been modified
correspondingly.

Some tests for spinor wavefunction normed inner products and the
updated lookup function for FermionBasis have been added.