Implement sign convention for fermionic operators
In this commit, an ordering convention for fermionic operators
is set for Fock states and for normal ordering. In generating
Fock states, we add particles first and then antiparticles;
in applying operators, operators should be normal-ordered, and then
we apply particle creation(/annihilation) operators before we apply
antiparticle creation(/annihilation) operators.

With this convention the sign ambiguity is removed.

Initial tests suggest that the potential matrix generated by this
implementation is symmetric, which is a promising first step. Still
need to remove the zero mode from basis states I think; operators
acting on the zero mode have already been excluded from the operators
in the interaction term.

Possible to-do: write code to normal-order a generic operator. If
we care about the contracted terms this is more difficult (i.e. the
delta functions are a bit tricky to deal with) but if we just need
the normal-ordered output this can be done quickly with Python
built-in sorting functions...

I have pseudocode for an implementation of this without keeping track
of the self-contractions; adding the delta function terms will be
harder.