Finish implementation of interaction term
This commit updates the zero mode spinor wavefunctions and
includes the various prefactors of the interaction term.

The computeHamiltonian method is updated so that calling runQED.py
with an appropriate radius and Emax (e.g. 1 6) will run and compute
the vacuum energy for these settings as well as the first few
eigenvalues.

Note that runQED now takes R and Emax as arguments rather than L
and Emax. This is easily changed if needed.

We fix the ordering of applying operators in FermionOperator. The
convention here is particle destruction, antiparticle destruction,
particle creation, antiparticle creation. Possibly more tests needed
to ensure all signs from anticommutations are correct here.

Modify __repr__ for FermionState to be more descriptive (i.e. to
specify particle and antiparticle occs).

Add a computeVacuumEnergy method to runQED to represent computing
the vacuum energy for one value of the coupling (compare
phi4_benchmark).