Fix signs in FermionOperator
Signs associated with operator ordering  in FermionOperator should
be correct now (i.e. the sum of filled slots runs up to index i,
not inclusive).

There was a missing factor of 1/L in the interaction terms, since
the operators (\psi^\dagger\psi)_k come with 1/sqrt(L) factors but
the overall coefficient is g^2/(2L*k^2)