Add normal-ordered (by hand) interaction term
Define a new method, generateOperators2, that (given nmax) generates
all the interaction term operators after normal-ordering.

Note that the method of producing the right momenta to loop over
(using a Numpy mask for the delta functions) may be more broadly
applicable to writing code to normal order a generic operator.

Writing out all the operators (and dropping the total delta function
terms) seems to have lifted the degeneracy I was noticing on the
eigenvalues.

Also, some annotations added to the renormalization code for phi4.