Update interaction term to remove zero operators
Due to a sign error, some of the terms which I originally wrote in
the interaction term are actually zero after imposing the delta
function constraints on momenta. The new interaction term is 24
operators (ranging over values of k1,k2,k3,k4).

There was also a coefficient error in the operator b_a_b_a_4
making the interaction term not symmetric (hermitian), but this has
been fixed.

A sign error resulted in left-movers and right-movers being switched,
but the net effect seems to have been minimal. I have updated the
spinor wavefunctions to fit the convention in my notes where the first
component is nonzero for *left-movers* (negative momenta) and the
second component is nonzero for right-movers (positive momenta).
I also update here the test functions which check the step function
behavior for the normed spinors.