Fix the interaction matrix
After much troubleshooting, Miranda and I agree on the matrix elements
for the 6-element basis (lmax=1). These should be totally right now;
we can see if this scales.

Fix an issue with the delta conditions (many pesky conventions). In
my sign convention, a delta means that k1 = -k2 because there is
a relative sign between a_k1 and a_{-k2}^\dagger.

To-do: remove operators with coefficient zero to speed up code.