Add axial charge calculation and saving (WIP!)
Add method to check axial charge of state

Add new scripts to divide the truncation into generation
of the matrices and diagonalization.
Diagonalization is fast and well-optimized, while matrix generation
takes most of the time.

This method is NOT validated yet! We can reproduce the analysis
at a fixed energy cutoff but trying to take lower cutoffs with a
larger file (e.g. Emax=4 for the Emax=16 data file) does not reproduce
the results from calculating Emax=4 directly. Not yet sure what the
issue is, but taking the submatrix (restricting to the matrix elements
with energies below some smaller cutoff) does not give the same matrix.

This could be an issue with the actual operators (although these were
validated against Miranda's version to good accuracy) or it could be
an issue with the implementation of saving. Hard to say for now, will
revisit this if we try to push to larger Emax.

Add new test methods and some plot PDFs made by hand.