Add DressedFermionState (including zero-mode)
New class: DressedFermionState
Includes the modification to the energy from the zero mode
of the gauge field. Has setter and getter methods for the state
of the zero-mode wavefunction (stored as a non-negative int).

Fix testQAxial to not automatically test up to Emax=13 (oops!) every
time we run unittest

Still to-do: update basis generation, check that old FermionState
operators are compatible with DressedFermionState. They should be,
since DressedFermionState extends FermionState. Implement operators
for the zero-mode part of the Fock state (including the weird
prefactors that count the number of fermions in the state).